Correction of issue #9
Te issue was in the creation of the new Login window: this new one was not linked to the recepteur (this is the only window to be connected directly and not through clientlocal, because it handles connexion information, it shall be separated.  Moreover, the account was deleted from the database. Now it just switch to the "loggedOFF" state.