Finished Userpath to editors
Finally made it:
Now child and parents can access their editors. For now it's the same editors, I don't think it'll be hard to add the differenciation.
Nothing after the editors.
For now, the drawing are created on every session, so this is not useful. But it's a proof of concept :smile: